+++ transport/http/outbound.go
@@ -42,8 +42,6 @@ import (
 	"github.com/opentracing/opentracing-go/ext"
 )
 
-var errOutboundNotStarted = errors.ErrOutboundNotStarted("http.Outbound")
-
 // this ensures the HTTP outbound implements both transport.Outbound interfaces
 var (
 	_ transport.UnaryOutbound              = (*Outbound)(nil)
@@ -75,6 +73,7 @@ func URLTemplate(template string) OutboundOption {
 // is called.
 func (t *Transport) NewOutbound(chooser peer.Chooser, opts ...OutboundOption) *Outbound {
 	o := &Outbound{
+		once:        sync.Once(),
 		chooser:     chooser,
 		urlTemplate: defaultURLTemplate,
 		tracer:      t.tracer,
@@ -170,10 +169,10 @@ func (o *Outbound) IsRunning() bool {
 
 // Call makes a HTTP request
 func (o *Outbound) Call(ctx context.Context, treq *transport.Request) (*transport.Response, error) {
-	if !o.IsRunning() {
-		// TODO replace with "panicInDebug"
-		return nil, errOutboundNotStarted
+	if err := o.once.WhenRunning(ctx); err != nil {
+		return nil, err
 	}
+
 	start := time.Now()
 	deadline, _ := ctx.Deadline()
 	ttl := deadline.Sub(start)
@@ -183,10 +182,10 @@ func (o *Outbound) Call(ctx context.Context, treq *transport.Request) (*transpor
 
 // CallOneway makes a oneway request
 func (o *Outbound) CallOneway(ctx context.Context, treq *transport.Request) (transport.Ack, error) {
-	if !o.IsRunning() {
-		// TODO replace with "panicInDebug"
-		return nil, errOutboundNotStarted
+	if err := o.once.WhenRunning(ctx); err != nil {
+		return nil, err
 	}
+
 	start := time.Now()
 	var ttl time.Duration