+++ transport/tchannel/inbound.go
@@ -26,26 +26,20 @@ import (
 	"go.uber.org/yarpc/internal/sync"
 )
 
-// ChannelInbound receives YARPC requests over TChannel.
-// It may be constructed using the NewInbound method on ChannelTransport.
-// If you have a YARPC peer.Chooser, use the unqualified tchannel.Transport
-// instead (instead of the tchannel.ChannelTransport).
-type ChannelInbound struct {
-	transport *ChannelTransport
-
-	once sync.LifecycleOnce
+// Inbound receives YARPC requests over TChannel. It may be constructed using
+// the NewInbound method on a tchannel.Transport.
+type Inbound struct {
+	once      sync.LifecycleOnce
+	transport *Transport
 }
 
 // NewInbound returns a new TChannel inbound backed by a shared TChannel
-// transport.  The returned ChannelInbound does not support peer.Chooser
-// and uses TChannel's own internal load balancing peer selection.
-// If you have a YARPC peer.Chooser, use the unqualified tchannel.NewInbound
-// instead.
+// transport.
 // There should only be one inbound for TChannel since all outbounds send the
 // listening port over non-ephemeral connections so a service can deduplicate
 // locally- and remotely-initiated persistent connections.
-func (t *ChannelTransport) NewInbound() *ChannelInbound {
-	return &ChannelInbound{
+func (t *Transport) NewInbound() *Inbound {
+	return &Inbound{
 		once:      sync.Once(),
 		transport: t,
 	}
@@ -54,44 +48,38 @@ func (t *ChannelTransport) NewInbound() *ChannelInbound {
 // SetRouter configures a router to handle incoming requests.
 // This satisfies the transport.Inbound interface, and would be called
 // by a dispatcher when it starts.
-func (i *ChannelInbound) SetRouter(router transport.Router) {
+func (i *Inbound) SetRouter(router transport.Router) {
 	i.transport.router = router
 }
 
-// Transports returns a slice containing the ChannelInbound's underlying
-// ChannelTransport.
-func (i *ChannelInbound) Transports() []transport.Transport {
+// Transports returns a slice containing the Inbound's underlying
+// Transport.
+func (i *Inbound) Transports() []transport.Transport {
 	return []transport.Transport{i.transport}
-}
-
-// Channel returns the underlying Channel for this Inbound.
-func (i *ChannelInbound) Channel() Channel {
-	return i.transport.ch
 }
 
 // Start starts this Inbound. Note that this does not start listening for
 // connections; that occurs when you start the underlying ChannelTransport is
 // started.
-func (i *ChannelInbound) Start() error {
+func (i *Inbound) Start() error {
 	return i.once.Start(nil)
 }
 
 // Stop stops the TChannel outbound. This currently does nothing.
-func (i *ChannelInbound) Stop() error {
+func (i *Inbound) Stop() error {
 	return i.once.Stop(nil)
 }
 
-// IsRunning returns whether the ChannelInbound is running.
-func (i *ChannelInbound) IsRunning() bool {
+// IsRunning returns whether the Inbound is running.
+func (i *Inbound) IsRunning() bool {
 	return i.once.IsRunning()
 }
 
 // Introspect returns the state of the inbound for introspection purposes.
-func (i *ChannelInbound) Introspect() introspection.InboundStatus {
-	c := i.transport.Channel()
+func (i *Inbound) Introspect() introspection.InboundStatus {
 	return introspection.InboundStatus{
 		Transport: "tchannel",
-		Endpoint:  i.transport.ListenAddr(),
-		State:     c.State().String(),
+		Endpoint:  i.transport.addr,
+		State:     i.transport.ch.State().String(),
 	}
 }